Step Selector: Stop unnecessarily rerendering data table with extra component layer
