+++ tensorboard/webapp/metrics/views/card_renderer/scalar_card_module.ts
@@ -28,6 +28,7 @@ import {TruncatedPathModule} from '../../../widgets/text/truncated_path_module';
 import {DataDownloadModule} from './data_download_module';
 import {ScalarCardComponent} from './scalar_card_component';
 import {ScalarCardContainer} from './scalar_card_container';
+import {ScalarCardDataTable} from './scalar_card_data_table';
 import {ScalarCardFobController} from './scalar_card_fob_controller';
 import {VisLinkedTimeSelectionWarningModule} from './vis_linked_time_selection_warning_module';
 
@@ -35,6 +36,7 @@ import {VisLinkedTimeSelectionWarningModule} from './vis_linked_time_selection_w
   declarations: [
     ScalarCardContainer,
     ScalarCardComponent,
+    ScalarCardDataTable,
     ScalarCardFobController,
   ],
   exports: [ScalarCardContainer],
